tools/vectorizer/requirements.txt
supabase>=2.3.0
google-genai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.26.0
tqdm>=4.66.0
